+++ src/main/java/org/lightcouch/CouchDbClientBase.java
@@ -1,4 +1,5 @@
 /*
+ * Copyright (C) 2018 indaba.es
  * Copyright (C) 2011 lightcouch.org
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -82,6 +83,7 @@ public abstract class CouchDbClientBase {
 	private Gson gson; 
 	private CouchDbContext context;
 	private CouchDbDesign design;
+	private Local local;
 	final HttpClient httpClient;
 	final HttpHost host;
 	
@@ -101,6 +103,7 @@ public abstract class CouchDbClientBase {
 		
 		this.context = new CouchDbContext(this, props); 
 		this.design = new CouchDbDesign(this);
+		this.local = new Local(this);
 	}
 	
 	// Client(s) provided implementation
@@ -145,6 +148,10 @@ public abstract class CouchDbClientBase {
 	 */
 	public View view(String viewId) {
 		return new View(this, viewId);
+	}
+	
+	public Local local() {
+	    return local;
 	}
 
 	/**