+++ src/test/java/org/lightcouch/tests/ReplicationTest.java
@@ -26,7 +26,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.UUID;
 
-import org.lightcouch.URIBuilder;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.junit.AfterClass;
@@ -34,12 +33,12 @@ import org.junit.Assume;
 import org.junit.BeforeClass;
 import org.junit.Test;
 import org.lightcouch.CouchDbClient;
-import org.lightcouch.CouchDbException;
 import org.lightcouch.DocumentConflictException;
 import org.lightcouch.ReplicationResult;
 import org.lightcouch.ReplicationResult.ReplicationHistory;
 import org.lightcouch.ReplicatorDocument;
 import org.lightcouch.Response;
+import org.lightcouch.URIBuilder;
 import org.lightcouch.ViewResult;
 
 public class ReplicationTest {