+++ CHANGES.md
@@ -1,4 +1,5 @@
 # 0.2.2 (UNRELEASED)
+- [NEW] Added explicit API for local document management.
 
 # 0.2.1 (21/02/2018)
 - [NEW] Added API for specifying a mango selector _changes operation